nixd/Server: split worker in a dedicated class
TODO: sync the draft store
