+++ nixd/lib/Support/CompletionHelper.cpp
@@ -1,27 +1,6 @@
-#include "nixd/Server/Server.h"
+#include "nixd/Support/CompletionHelper.h"
+#include "nixd/Nix/Eval.h"
 
-#include <nix/eval-inline.hh>
-#include <nix/shared.hh>
-
-namespace nix {
-
-// Copy-paste from nix source code, do not know why it is inlined.
-
-inline void EvalState::evalAttrs(Env &env, Expr *e, Value &v, const PosIdx pos,
-                                 std::string_view errorCtx) {
-  try {
-    e->eval(*this, env, v);
-    if (v.type() != nAttrs)
-      error("value is %1% while a set was expected", showType(v))
-          .withFrame(env, *e)
-          .debugThrow<TypeError>();
-  } catch (Error &e) {
-    e.addTrace(positions[pos], errorCtx);
-    throw;
-  }
-}
-
-} // namespace nix
 namespace nixd {
 
 void CompletionHelper::fromStaticEnv(const nix::SymbolTable &STable,
@@ -59,15 +38,6 @@ void CompletionHelper::fromEnv(nix::EvalState &State, nix::Env &NixEnv,
   }
   if (NixEnv.up)
     fromEnv(State, *NixEnv.up, Items);
-}
-
-void Server::initWorker() {
-  assert(Role == ServerRole::Controller &&
-         "Must switch from controller's fork!");
-  nix::initNix();
-  nix::initLibStore();
-  nix::initPlugins();
-  nix::initGC();
 }
 
 } // namespace nixd
